Upgrade the example to RN 0.69
Only tested on iOS for now.

I updated TUSKit to the latest 2.x version, but `response?.errorCode` does not seem to exist anymore?

Also when adding files `file:/` is added to the start of the path somehow, even if it already starts with `file://` as the `react-native-image-picker` URLs do, so I removed it in the example.